monitor/requirements.txt
pyserial==3.5
argparse==1.4.0
